+++ packages/server/test/integration/websockets_spec.js
@@ -39,7 +39,7 @@ describe('Web Sockets', () => {
         SocketCtor: SocketE2E,
         createRoutes,
         specsStore: new SpecsStore({}, 'e2e'),
-        projectType: 'e2e',
+        testingType: 'e2e',
       })
       .then(async () => {
         const automationStub = {