+++ npm/vue/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@cypress/vue-v3.0.3](https://github.com/cypress-io/cypress/compare/@cypress/vue-v3.0.2...@cypress/vue-v3.0.3) (2021-07-31)
+
+
+### Bug Fixes
+
+* vue 3 types, beta suffix & component name ([#17508](https://github.com/cypress-io/cypress/issues/17508)) ([b4733a6](https://github.com/cypress-io/cypress/commit/b4733a6179577b0447af9bcdcf417713d5fb38b9))
+
+# [@cypress/vue-v3.0.0-beta.4](https://github.com/cypress-io/cypress/compare/@cypress/vue-v3.0.0-beta.3...@cypress/vue-v3.0.0-beta.4) (2021-07-27)
+
+
+### Bug Fixes
+
+* do not register CT specific event in e2e mode ([5836203](https://github.com/cypress-io/cypress/commit/58362037fd231c6fdc587e422fc139bf7546ac2d))
+
 # [@cypress/vue-v3.0.0-beta.3](https://github.com/cypress-io/cypress/compare/@cypress/vue-v3.0.0-beta.2...@cypress/vue-v3.0.0-beta.3) (2021-06-24)